support Stack.Screen, etc, in RSC
